implement more wrappers

+++ src/main/scala/engine/parser/wrapper/OperandWrapper.scala
@@ -8,6 +8,6 @@ case class OperandWrapper() {
 }
 
 object OperandWrapper {
-  def apply(num: Int): OperandWrapper = {OperandWrapper()}
+  def apply(num: Double): OperandWrapper = {OperandWrapper()}
   def apply(variable: Char): OperandWrapper = {OperandWrapper()}
 }
