+++ src/main/scala/engine/parser/wrapper/ParamWrapper.scala
@@ -3,6 +3,6 @@ package engine.parser.wrapper
 /**
   * Created by et on 22.03.17.
   */
-case class ParamWrapper(value: Int, unit: Option[String]) {
+case class ParamWrapper(value: Double, unit: Option[String]) {
 
 }
